fix: widen CodeLoc line/column from u16 to u32 to prevent overflow on long lines
CodeLoc stored line and column as u16, which panics at 65,536.
Parsing a single-line string literal >65K chars (e.g. from
SessionManager::set_variable with a large repr() value) triggered
"Column number overflow" in CodeLoc::new.

Widen both fields to u32 (~4.3B max) across all downstream types:
CodeLoc, Generator.saved_lineno, suspend_generator_frame,
first_location_after_line, warning_source_line, PyFrame, and
new_generator_frame_proxy.

+++ crates/ouros/src/bytecode/code.rs
@@ -155,7 +155,7 @@ impl Code {
     ///
     /// Useful for locating the first executable line inside a compound statement body.
     #[must_use]
-    pub(crate) fn first_location_after_line(&self, line: u16) -> Option<CodeRange> {
+    pub(crate) fn first_location_after_line(&self, line: u32) -> Option<CodeRange> {
         let mut best: Option<CodeRange> = None;
         for entry in &self.location_table {
             let range = entry.range();
